+++ backend/core/js/tiny_mce/plugins/dextrose_videoembed/langs/de.js
@@ -1,7 +1,0 @@
-tinyMCE.addI18n("de.dextrose_videoembed",{
-	delta_width : 0,
-	delta_height: 0,
-	video: 'Video hinzufügen',
-	title: 'Video hinzufügen',
-	description: 'Fügen Sie Ihre Video-Code'
-});