+++ backend/core/js/tiny_mce/plugins/dextrose_videoembed/langs/tr.js
@@ -1,7 +1,0 @@
-tinyMCE.addI18n("tr.dextrose_videoembed",{
-	delta_width : 0,
-	delta_height: 0,
-	video: 'Video ekle',
-	title: 'Video ekle',
-	description: 'Video kodu yapıştırın'
-});