+++ backend/modules/extensions/engine/model.php
@@ -744,8 +744,12 @@ class BackendExtensionsModel
 		$installer->install();
 
 		// save the warnings in session for later use
-		$warnings = array('module' => $module, 'warnings' => $installer->getWarnings());
-		SpoonSession::set('installer_warnings', $warnings);
+		if($installer->getWarnings())
+		{
+			$warnings = SpoonSession::exists('installer_warnings') ? SpoonSession::get('installer_warnings') : array();
+			$warnings = array_merge($warnings, array('module' => $module, 'warnings' => $installer->getWarnings()));
+			SpoonSession::set('installer_warnings', $warnings);
+		}
 
 		// clear the cache so locale (and so much more) gets rebuilt
 		self::clearCache();