+++ frontend/themes/triton/core/layout/editor_templates/templates.js
@@ -1,0 +1,8 @@
+[
+	{
+		title: 'A tribute to Fork',
+		image: '/frontend/themes/triton/core/layout/editor_templates/images/icon.png',
+		description: '',
+		html: '<p><img class="alignLeft" src="/frontend/themes/triton/core/layout/editor_templates/images/icon.png" height="70" width="100" />This theme was made for <a href="http://www.fork-cms.com">Fork</a>.</p>'
+	}
+]